requirements.txt
PyYAML==5.3.1
Flask==2.1.2
requests==2.22.0